1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int t;
+	size_t t;
 
 	if (array == 0 || size == 0)
 		return;
